+++ src/main/java/com/github/sgoeschl/freemarker/cli/tools/excel/ExcelTool.java
@@ -16,6 +16,7 @@
  */
 package com.github.sgoeschl.freemarker.cli.tools.excel;
 
+import com.github.sgoeschl.freemarker.cli.impl.CloseableReaper;
 import com.github.sgoeschl.freemarker.cli.model.Document;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.DataFormatter;
@@ -24,6 +25,7 @@ import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
 
+import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -32,11 +34,24 @@ import java.util.List;
 
 import static org.apache.poi.ss.usermodel.Row.MissingCellPolicy.CREATE_NULL_AS_BLANK;
 
-public class ExcelTool {
+public class ExcelTool implements Closeable {
+
+    private final CloseableReaper closeableReaper;
+
+    public ExcelTool() {
+        this.closeableReaper = new CloseableReaper();
+    }
+
+    @Override
+    public void close() {
+        closeableReaper.close();
+    }
 
     public Workbook parse(Document document) {
         try (InputStream is = document.getInputStream()) {
-            return WorkbookFactory.create(is);
+            final Workbook workbook = WorkbookFactory.create(is);
+            closeableReaper.add(workbook);
+            return workbook;
         } catch (IOException e) {
             throw new RuntimeException("Failed to parse Ecxel document: " + document, e);
         }
@@ -57,16 +72,19 @@ public class ExcelTool {
 
         while (iterator.hasNext()) {
             final Row row = iterator.next();
-            final List<String> columnValues = new ArrayList<>();
-            for (int columnIndex = 0; columnIndex < row.getLastCellNum(); columnIndex++) {
-                final Cell cell = row.getCell(columnIndex, CREATE_NULL_AS_BLANK);
-                final String formatedCellValue = dataFormatter.formatCellValue(cell).trim();
-                columnValues.add(formatedCellValue);
-            }
-            result.add(columnValues);
+            result.add(toColumns(row, dataFormatter));
         }
 
         return result;
     }
 
+    private static List<String> toColumns(Row row, DataFormatter dataFormatter) {
+        final List<String> columnValues = new ArrayList<>();
+        for (int columnIndex = 0; columnIndex < row.getLastCellNum(); columnIndex++) {
+            final Cell cell = row.getCell(columnIndex, CREATE_NULL_AS_BLANK);
+            final String formatedCellValue = dataFormatter.formatCellValue(cell).trim();
+            columnValues.add(formatedCellValue);
+        }
+        return columnValues;
+    }
 }